refactor: Move tab list to each domain directory
